#318|mridul, sandeep| Changing program code to program id and changing id type to Long from int

+++ modules/core/src/main/java/org/openlmis/core/domain/User.java
@@ -7,11 +7,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class User {
 
-    private int id;
+    private Long id;
     private String userName;
     private String password;
     private String role;
-    private Integer facilityId;
+    private Long facilityId;
 
     public User(String userName, String password) {
         this.userName = userName;
