+++ modules/core/src/main/java/org/openlmis/core/repository/ProductRepository.java
@@ -62,7 +62,7 @@ public class ProductRepository {
     }
   }
 
-  public List<Product> getByFacilityAndProgram(int facilityId, String programCode) {
+  public List<Product> getByFacilityAndProgram(Long facilityId, String programCode) {
         return mapper.getFullSupplyProductsByFacilityAndProgram(facilityId, programCode);
     }
