+++ modules/requisition/src/main/java/org/openlmis/rnr/mapper/RequisitionGroupMapper.java
@@ -14,7 +14,7 @@ public interface RequisitionGroupMapper {
     @Select("INSERT INTO requisition_group(code,name,description,level_id,head_facility_id,parent_id,active,modified_by) " +
             "values (#{code},#{name},#{description},#{levelId}, #{headFacility.id} ,#{parent.id},#{active},#{modifiedBy}) returning id")
     @Options(useGeneratedKeys = true)
-    Integer insert(RequisitionGroup requisitionGroup);
+    Long insert(RequisitionGroup requisitionGroup);
 
     @Select("SELECT rg.id, rg.code, rg.name, rg.description, rg.level_id, rg.head_facility_id, " +
             "rg.parent_id, rg.active, rg.modified_by, rg.modified_date FROM " +
@@ -31,5 +31,5 @@ public interface RequisitionGroupMapper {
             @Result(property = "modifiedBy", column = "modified_by"),
             @Result(property = "modifiedDate", column = "modified_date")
     })
-    RequisitionGroup getRequisitionGroupById(int id);
+    RequisitionGroup getRequisitionGroupById(Long id);
 }
